LayoutTests/platform/wpe/http/tests/websocket/tests/hybi/deflate-frame-invalid-parameter-expected.txt

Test whether WebSocket rejects invalid deflate-frame parameters.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Testing parameter: "x-foo"
onclose() was called.
PASS closeEvent.wasClean is false
Testing parameter: "max_window_bits=7"
onclose() was called.
PASS closeEvent.wasClean is false
Testing parameter: "max_window_bits=16"
onclose() was called.
PASS closeEvent.wasClean is false
Testing parameter: "no_context_takeover=foo"
onclose() was called.
PASS closeEvent.wasClean is false
Testing parameter: "max_window_bits=8; no_context_takeover; x-foo"
onclose() was called.
PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
